Get rid of Authenticator interface
